+++ test/api/generated-parser-api.spec.js
@@ -95,6 +95,7 @@ describe("generated parser API", function() {
                 type: "rule.enter",
                 rule: "start",
                 location: {
+                  source: undefined,
                   start: { offset: 0, line: 1, column: 1 },
                   end: { offset: 0, line: 1, column: 1 }
                 }
@@ -103,6 +104,7 @@ describe("generated parser API", function() {
                 type: "rule.enter",
                 rule: "a",
                 location: {
+                  source: undefined,
                   start: { offset: 0, line: 1, column: 1 },
                   end: { offset: 0, line: 1, column: 1 }
                 }
@@ -111,6 +113,7 @@ describe("generated parser API", function() {
                 type: "rule.fail",
                 rule: "a",
                 location: {
+                  source: undefined,
                   start: { offset: 0, line: 1, column: 1 },
                   end: { offset: 0, line: 1, column: 1 }
                 }
@@ -119,6 +122,7 @@ describe("generated parser API", function() {
                 type: "rule.enter",
                 rule: "b",
                 location: {
+                  source: undefined,
                   start: { offset: 0, line: 1, column: 1 },
                   end: { offset: 0, line: 1, column: 1 }
                 }
@@ -128,6 +132,7 @@ describe("generated parser API", function() {
                 rule: "b",
                 result: "b",
                 location: {
+                  source: undefined,
                   start: { offset: 0, line: 1, column: 1 },
                   end: { offset: 1, line: 1, column: 2 }
                 }
@@ -137,6 +142,7 @@ describe("generated parser API", function() {
                 rule: "start",
                 result: "b",
                 location: {
+                  source: undefined,
                   start: { offset: 0, line: 1, column: 1 },
                   end: { offset: 1, line: 1, column: 2 }
                 }